backend/requirements.txt
# ── EduRAG Backend Dependencies ──────────────────────────────────

# Core Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1

# Database
sqlalchemy==2.0.36
alembic==1.14.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20

# PDF Processing
PyMuPDF==1.25.3

# OCR (optional - for scanned PDFs)
pytesseract==0.3.13
Pillow==11.1.0

# AI / ML
sentence-transformers==3.3.1
torch==2.5.1
transformers==4.47.1
huggingface-hub==0.27.1

# LLM Runtime — Install separately (see README):
#   GPU:  pip install llama-cpp-python --extra-index-url https://abetlen.github.io/llama-cpp-python/whl/cu124
#   CPU:  pip install llama-cpp-python --extra-index-url https://abetlen.github.io/llama-cpp-python/whl/cpu
# llama-cpp-python==0.3.4

# Vector Database
faiss-cpu==1.9.0.post1
# For GPU: faiss-gpu==1.7.2

numpy==2.0.2  # Python 3.9 compatible (2.2+ requires Python 3.10+)

# Email Validation
email-validator==2.2.0

# Logging
coloredlogs==15.0.1
